backend/requirements.txt
flask==2.3.2
flask-cors==4.0.1
pymongo==4.6.1
python-dotenv==1.0.1
google-generativeai==0.8.3
